agent: remove unused code and move some files out
