--- internal/agent/time.go
@@ -1,10 +1,0 @@
-package agent
-
-import (
-	"time"
-)
-
-// Now returns a timestamp in our nanoseconds default format
-func Now() int64 {
-	return time.Now().UnixNano()
-}
