+++ internal/api/limited_reader.go
@@ -1,4 +1,4 @@
-package agent
+package api
 
 import (
 	"errors"
