+++ internal/api/limited_reader_test.go
@@ -1,4 +1,4 @@
-package agent
+package api
 
 import (
 	"bytes"
